\name{disconnect.from.db}
\alias{disconnect.from.db}
\title{Disconnect from a database}
\usage{
  disconnect.from.db(conn = default.connection)
}
\arguments{
  \item{con}{The connection object to close. Defaults to
  default.connection}
}
\description{
  This function will close an active database connection.
}

